Added a close function
The onclick function closes the menu now which no longer requires the
user to click the hamburger icon. The stopPropagation() prevents the
menu from immediately closing when opened.

Random bug is the menu closes on its own. This doesn't seem to happen
unless the action is cycled repeatedly.

The original code was pulled from CodePen which is commented in the
HTML.